ui: Release canary
Commands:
    $ git fetch origin
    $ git checkout -B ui-canary -t origin/ui-canary
    $ git merge --strategy=ours origin/main
    $ git diff --binary origin/main | git apply --reverse --index
    $ git commit --amend

    Check end state is correct:
    $ git diff ui-canary origin/main
    [no output]
    $ git rev-list --count ui-canary..origin/main
    0

+++ include/perfetto/ext/base/small_set.h
@@ -49,7 +49,9 @@ class SmallSet {
   }
 
   const_iterator begin() const { return arr_.cbegin(); }
-  const_iterator end() const { return arr_.cbegin() + filled_; }
+  const_iterator end() const {
+    return arr_.cbegin() + static_cast<ssize_t>(filled_);
+  }
   size_t size() const { return filled_; }
 
  private:
